requirements.txt
streamlit
PyPDF2
langchain
faiss-cpu
python-dotenv
google-generativeai
langchain-google-genai